+++ src/components/keypad/Keypad.tsx
@@ -1,4 +1,4 @@
-import { TEAMS, TEAM_TRANSLATION } from "common/constants";
+import { TEAMS, TEAM_TRANSLATION } from "game/constants";
 import { useGameContext } from "game/context/GameContext";
 import { Team } from "game/reducer/types";
 import React, { FC, useState } from "react";
