+++ src/components/scoreboard/foot/Foot.tsx
@@ -1,4 +1,4 @@
-import { TEAMS } from "common/constants";
+import { TEAMS } from "game/constants";
 import { useGameContext } from "game/context/GameContext";
 import React, { FC } from "react";
 import styles from "../Scoreboard.module.css";
